Autoupdate of 2 submodules
*tools/adt/idea updated to 714d486c3ee3d7e0d56990da739d1a663439cae2
Make isBuilding not require UiThread

in ProjectBuildStatusManager. The requirement is there to ensure that
builds can not start/finish causing a race condition. However this can
easily introduce dead-locks. For the potential use cases that care about
having a consistent view of the build status, we should probably have a
separate mechanism like runWhenNotBuilding or they should guarantee
themselves the check by calling isBuilding within the UI thread.

An example of these dead-locks would be to call
ComposePreviewRepresentation.status() from a background thread. When
this happens from an action, IntelliJ will automatically hold the read
lock for the action. This causes that, if we pass it to the UI thread
and the UI thread is doing something that requires a Write action, the
UI thread will never complete the action

Bug: N/A
Test: Existing
Change-Id: Ifbd54c094a58bf65327019800d8cd8d83b5fe75e

*tools/vendor/google3/blaze updated to 715e88581c883f0efbfa5e6d2f69d389f63f5586
Remove UiThread from isBuilding in ProjectSystemBuildManager

Test: Existing
Bug: N/A
Change-Id: I0a33d436109a5f80236c8414b371a0fe20ff1a66