Fix Live Edit indicator for non-debuggable apps
AGP is setting the debuggability of the APK to false when we launch
with "Profile with low overhead". This change is not reflected in
the AndroidModel. And since parsing the APK is too big of a change
at the current time, we are just going to work around the Profiler
for now, and will follow up with a proper refactor to use the
manifest as source of truth (as tracked by the TODO).

TODO: 281742972
Fixes: 280460949
Test: run with "Profile with low overhead" or in Release
Change-Id: I4f572c1e25964775d269ce762fc328c9835050dc