Keep assistant usable when metrics can't get the project ID
Sometimes when getting the project ID we raise an error instead. This
caused the "run all filters on open" behavior to get stuck and load
forever.

Metrics issues shouldn't prevent usage of the assistant. Thus, we now
catch this error, log it as a warning (to avoid the red blinker or
polluting crash), and pretend as if we got a null project ID instead,
which was already handled. This allows the filter to return normally,
rendering the assistant usable, and will be visible in metrics as coming
from the "unknown" project ID.

Fixes: 282985770
Test: N/A
Change-Id: Iab2430617039f122eba0dcc826c9e2dff2bbc6b1