[PSD] Clear the tree's cell renderer on dispose
The PsContext should be disposed when the ProjectStructureConfigurable
is disposing its UI resources, which ends up removing all the actual
Configurables from the ProjectStructureConfigurable, but does not
clear all UI state from the ProjectStructureConfigurable itself.  We
will end up creating new trees if we ever return to the PSD, as we
will construct whole new Configurables.

Also clear the various fields holding onto bits of stale UI from
ProjectStructureConfigurable, because that seems reasonable.

Bug: 279766006
Test: ?
Change-Id: I1f4a69990adae28c4576e92aa86cbb40cc5634bf