Autoupdate of 2 submodules
*tools/analytics-library updated to a04ed0e80ad69843713ef2a65205f65f4d96615f
Add LINT_RESERVED_MEMORY_PER_TASK to analysis_enums.proto

Bug: 197322928
Test: existing
Change-Id: I7d52a7641309e42c403de507f8fdd03017c58deb

*tools/base updated to 212bae07ef1463bf0a530dde04dbf428c22ac1ed
Decrease the default amount of memory earmarked for lint tasks

The amount of memory earmarked for lint tasks was recently increased
from 512mb to 1024mb. This change allows users to set how much memory to
reserve per lint task and decreases the default back to 512mb.

The motivation for this decrease is the plan to move to per component
lint analysis.

Bug: 197322928
Test: LintParallelBuildServiceTest
Change-Id: I7cc608bc12d5490509770bfd425cf631e5d6e835