Fix Bug When Target File Doesn't Exist
`VirtualFileWrapper.getVirtualFile() returns null if the file doesn't
exist yet and `FileSaverDialog` does not create the file.

This causes `SaveLogcatAction` to only work if the target file already
existed. While there is a version of
`VirtualFileWrapper.getVirtualFile(boolean createIfNotExist)` that would
ensure the file exists, I chose to use `VirtualFileWrapper.getFile()`
instead and resolve the `VirtualFile` later because it saves the extra
step of creating the file explicitly.

The test was passing because the fake `FileSaverDialog` was actually
creating a file unlike the prod version which does not.

Bug: NA
Test: Fixed bad test
Change-Id: Iaaaae78cf252d767287fdaafc6d48367c31b47e3