[FTL GMD] create test runner class
Moving out the testing implementation from the TestLabBuildService and
moving it to a designated TestRunner class to allow for easier unit
testing and code management.

The file operations of the TestRunner as well as the credentials file
operations in the TestLabBuildService still need to be separated out.

Test: manual for now. Need to move out the file operations to make unit
tests for the TestRunner class.

Bug: 278583488
Change-Id: I5adb7869581ecfae87500a9ab44356b2462770f7