syscalls/statx06: use a fine-grained timestamp for the second time fetch
I have a patchset in progress to change the kernel to sometimes use
fine-grained timestamps for the mtime/ctime. With this, the statx06 test
sometimes fails.

Change the test to grab a fine-grained timestamp for the "after" value,
which fixes the issue.

Link: https://lore.kernel.org/ltp/20230518113216.126233-1-jlayton@kernel.org/
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Yang Xu <xuyang2018.jy@fujitsu.com>
[ pvorel: Although patchset [1] has not been merged, we agreed that
the change is backward compatible anyway. ]
Signed-off-by: Jeff Layton <jlayton@kernel.org>

[1] https://lore.kernel.org/lkml/20230411143702.64495-1-jlayton@kernel.org/

(cherry picked from commit 604701dc19c5499403ae5afe48103b8e2872952b)
Bug: 300322453
Bug: 300116501
Test: atest vts_ltp_test_x86:syscalls.statx06_32bit
Change-Id: I5a82c25878782daf6cef3dcc87e9b2b2a41b4bcf
Signed-off-by: Edward Liaw <edliaw@google.com>