Process tools attributes in privacy sandbox sdk manifests
Previously the same settings were being used for fused library and privacy
sandbox sdk plugins, which was not deliberate. (This change forks that task
worker action to make it simpler for them to diverge going forward)

Remove (somewhat broken) support for propogating permissions to the consuming
app, at least for now it's not needed.

Test: Included
Bug: 283210991
Bug: 283763669
Change-Id: Id5968e0f0114bf6666b913e1dc75777069d3e16e