Autoupdate of 3 submodules
*tools/adt/idea updated to baf1f065cdecbab26d88af9dbe6559c1bd9c839c
Delete deprecated AndroidModuleModel and update existing usages

This model was already replaced in favor of using AndroidModel as an
abstraction for the supported build systems or GradleAndroidModel for
Gradle specific.

Fixes: 22596984
Test: Existing
Change-Id: I5209f2d1e445929ffad6064ec3dd29129ec36a02

*tools/studio/google/cloud/testing updated to 3cac5738b6efae72845f97613fb5362d3cc0763a
Delete deprecated AndroidModuleModel and update existing usages

This model was already replaced in favor of using AndroidModel as an
abstraction for the supported build systems or GradleAndroidModel for
Gradle specific.

Fixes: 22596984
Test: Existing
Change-Id: Ifea8003383a943ca0b4f04f04401ae6692a659d1

*tools/vendor/google updated to 349321d58a7e6ea4c3d169b2729407ad09bd8d07
Delete deprecated AndroidModuleModel and update existing usages

This model was already replaced in favor of using AndroidModel as an
abstraction for the supported build systems or GradleAndroidModel for
Gradle specific.

Fixes: 22596984
Test: Existing
Change-Id: I48c2bbfd1c9a7631f483d283eda4ba7a04b2c1a8