Autoupdate of 2 submodules
*tools/adt/idea updated to 990a3dfd8d51829755a7c6b524220a99c724c034
Use source version of samplesindex api lib

Also delete some obsolete stuff.

Test: updated baseline
Bug: 283741253
Change-Id: I1036f1fdae7e1b06eed259104e6a69a7fdc53ddc

*tools/studio/google/cloud/tools updated to adba71434bcd05dda03cd2bc9f5e120d79e21a84
Use source version of samplesindex lib

...and modify it to accept the updated google-api-client.
The source version was created by decompiling the version checked in in
aosp/Ic075bb23505bda018e108b20d1a33b7261a14920.
I haven't yet been able to completely determine the origin of the original jar. However, I think this is what happened: the original code was in google3 before it was moved to gob and at that point it would have been possible to run the normal rest api generator on it--and the problematic check is exactly what you get from generating a jar in that way, as we've seen in other recently-updated jars.
Now that the code is moved to gob (and has evolved for nine years), there's no longer an easy way to generate such an api jar. So, at least pending further investigation, it seems reasonable to use the decompiled generated code directly (including adding the same license to the code as it was originally distributed under).

Also deleting some obsolete stuff.

Test: none for now. Adding one can be tracked by b/283741253
Bug: 283741253
Change-Id: I0a81329a4ebb1e21c37112e6d60cd602ceadb383