Invalidate binding caches on index changes
LocalResourceRepository and BindingXmlIndex may be updated at
different times, so we must incorporate both into the modification
count used by LayoutBindingModuleCache.

This is especially important in IntelliJ 2023.1 because smart mode may
be entered before all project files are indexed
(https://youtrack.jetbrains.com/issue/IJPL-50).

It is infeasible to write a regression test because b/283753328
involves a race with UnindexedFilesScanner.

Fixes: 283753328
Test: infeasible (see above)
Change-Id: I58acf8967b7f2675181892156b74da40f9d55d8c