Autoupdate of 2 submodules
*prebuilts/tools updated to 0b046cc674a04720120b45b8bf3fccd64b980805
Add origin information for downloaded artifacts

Bug: 280336914
Test: n/a
Change-Id: I88f66a790a2c06e8229d8d607e709dfd3db64899

*tools/base updated to 399fd51c2f59911f65f5e2551bb395f1bf55a70f
Capture the origin info for downloaded artifacts

This change adds the basic infrastructure for recording origin
information for downloaded Maven artifacts.

With this change, when the repository generator tool downloads a
Maven artifact, it also writes an `origin.json` file in the same
artifact folder. This file contains the repository information
for each file in the artifact (e.g., pom, jar).

Bug: 280336914
Test: Updated LocalMavenRepositoryTest
Change-Id: I93a24825b4857aeeffacb42bede4e441673098e1