[Insights] Refactor code around AppInsightsExternalAnnotator
This is to prepare for the following feature work.

Code in CrashToLineNaiveMapper is consolidated in
the controller. So the corresponding standalone
test can be removed.

How we generate annotations are re-arranged --
there's no real change, just it's for preparing
for the following feature work. Also the corresponding
unit tests are re-written to cover more cases in
the future.

Bug: 281702889
Test: updated
Change-Id: Ic15eaa5668cca99f91d3855f4add64cc76dc3099