Implement getPreferredSize() for PropertyComboBox
The change: If1fd2a0a introduced a problem where a PropertyComboBox
would show up with a squeezed height if used outside of a property
table. Implement getPreferredSize() on the JPanel that holds the
comboBox and the label used for hover expansions.

Also: add a few defaults to FakeComboBoxUi to avoid NPEs when calling
getPreferredSize() on a comboBox in tests using FakeComboBoxUi.

Fixes: 284255907
Test: Added unit test
Change-Id: I60c7e98694ad758a4f00d9fdd37b99df486763f4