Autoupdate of 2 submodules
*tools/adt/idea updated to 5975c07cdfb8f98d376969d434e8a8caaf6a0b60
Adding buildConfig to module deps

The IDE cannot read buildConfig
from bytecode because it is from a jar.
Add the jar as a module dependency in
this case.

Bug: 280000337
Test: GradleSyncProjectComparisonTest, SyncedProjectTest
Change-Id: I809993ff3e553ba5653036d58bfd19d5cfdb2dbb

*tools/base updated to 6de1c3cf058abac444320858ef7197d949048f70
Adding buildConfig to classesFolder

AS does not resolve references for
BuildConfig when enableBuildConfigAsBytecode
is enabled. Add to classesFolder so it can be
added as module dependency in the IDE.

Bug: 280000337
Test: AppModelTest
Change-Id: I0f1909024871d68cc96b72881f0f59e24dafb1e6