Merge commit 'de0f89f2361e19da5eaeec38e50c4ac0141272c8' into aosp/master
Conflicts in a number of files:
* Api.java - both branches added `getClasses()`, albeit in different
  places in the file. That caused conflicts in the imports. Fixed by
  keeping the `getClasses()` from aosp/master as that returns an
  immutable list which is safer.
* ApiClass.java - conflicts resolved automatically by mergetool.
* ApiGenerator.java - conflict caused by replacing a call to
  `createApiLevelsXml(...)` which had different number of arguments in
  each branch with a call to `ApiJsonPrinter.print(...)`. Resolved by
  just taking the new JSON code.
* ApiGeneratorTest.kt - conflict caused by different version numbers in
  `<api version="..">` that is replaced with JSON. Resolved by just
  taking the new JSON output.

Also had to add `gson-prebuilt-jar-2.9.1` to `Android.bp`. A prebuilt
has to be used as `metalava` is used by Bazel and Bazel does not yet
supporting building arbitrary java libraries from sources. So, while
Bazel can build `metalava` it cannot yet build `gson` which is built
from the sources in `external/gson`. The version is slightly newer than
what is used in the `build.gradle.kts` file but it compiles.

Bug: 281035030
Test: ./gradlew
      m checkapi
Change-Id: I6f1bb7407a928df67de16e875c8505257d8db85a