[BA] add windows defender check warning
WindowsDefenderCheckService is responsible for providing windows
defender warning data to Build Analyzer UI. It is only requested from
the UI when new report is being opened. Since it is general warning
about windows defender state it does not depend on the build data
in any way, so there is no point to include this on Build Analyzer
backend.

When warning needs to be shown, corresponding page with the details
is generated in the warnings section.
Since no logic of this service is depending directly on windows, it is
tested in current setup on linux nodes using the mocked output from the
WindowsDefenderChecker - a service provided by the platform actually
doing all the hard work for the checks and configuration update.

Bug: 285867050
Test: Added WindowsDefenderWarningUITest.kt and
 WindowsDefenderCheckServiceTest.kt
Change-Id: I5174969a5f96e779b14d3f9cec1aed5bacfd2e5d