Use auto-import index for reference lookup
This CL makes a series of related changes to the smart-import
feature of Studio Bot code responses.

1. We now use the Maven auto-import index, which has been updated
to index more packages, and top-level functions from a select few.

2. Most of our hardcoded dataset is no longer needed. What we still need
to catalog ourselves are import packages missing from the index,
extension functions from all packages, and top-level functions from
some packages that aren't currently covered by the index. So the dataset
has been broken up into these pieces.

3. Since now we can tell if something is an extension function or not,
we can return to qualifying top-level functions in-place. For some time
we have been importing all names that start with lower-case letters, as
a safeguard against accidentally qualifying an extension function in-
-place. But besides catching many top-level functions, this approach
also simply does not work, since many extension functions in AndroidX
start with a capitalized letter. We no longer need such heuristics,
because we know that there are no extension functions in the Maven
index, and so we only need to import names which are found in our
extension function catalog. We also keep all imports that were
originally in the code snippet, because we can't tell if they refer
to an extension function and we don't want to re-introduce imprecise
heuristics.

4. KotlinParserTest and KotlinActionProcessorTest have been moved
to the sml-android test module,since parsing now needs access to
the Android plugin for the Maven index.

Bug: 274832437
Test: KotlinParserTest, SimpleNameRegistryTest
Change-Id: I27fcfa505908eabee2be20e11f74f8dad99bfccb