Address flaky test
This test is passing 100% when I've run it multiple hundreds of times
locally, both in IDEA and via Bazel. Reactivating this test accordingly,
although I'll check up on the postsubmits to ensure it's not failing in
that environment.

(If it is, I'll investigate a similar fix to ag/23474505; I just want to
avoid adding extra logic to the test if it's not needed.)

Fixes: 263898646
Test: AndroidLayoutDomTest.testJavaCreateResourceFromUsage
Change-Id: Ic4d6d649506d9f304ce2d0a758921249e566b413