Don't modify code from editor actions
We shouldn't do any post-processing on code snippets which are
returned from editor code actions like "Document code". These actions
pass along code from the user's project and expect a version of the
same code as the response. So it's not right to treat this code as
being generated by the model and so possibly needing fixups.
Instead we should leave it as-is so that the actions don't make
unexpected changes to the user's code.

Test: testDocumentQuery
Fixes: 284002839
Change-Id: Ic7d227a54887a30562765a7903e7ae8570f0763f