Refactor Kotlin code actions
This refactoring was spurred by b/280556462, which pointed out that comment code snippets weren't being inserted by the "Insert at Cursor" action. Comments are unique in that they can be inserted almost anywhere in a file without breaking syntax rules. And so inserting them shouldn't involve any PSI inspection, or treating them as elements in the PSI tree at all. It makes more sense to just paste them as text at the cursor.

The code action processors currently operate *only* on the PSI tree, and insertion takes the form of inserting snippets, parsed as PsiElements, into the tree. In order to support inserting comments as text, this would need to change. Also, this reliance on PSI manipulation helps in some cases, like adding members to a class, but it has been clearly the wrong approach for adding blocks of code right at the cursor, e.g. inside a function body, and involves confusing interactions with whitespace and newlines.

So the first major change to the KotlinActionProcessor is that it now operates chiefly on text offsets instead of PSI elements. The closest PSI element to the target offset is still computed, but now instead of being used an anchor point for insertion, it's main purpose is determining where in the PSI tree the cursor is located. This change required that large parts of code be totally rewritten, and so along with this change, many names and signatures were changed and organization was generally improved.

In order to test this new version of KotlinActionProcessor, KotlinActionProcessorTest had to be refactored as well. It is now much easier to write tests, and the tests are more readable and concise. See further instructions in the file.

While refactoring the tests, it became clear that there were some coverage gaps. This CL also closes some of those gaps by running the tests on a background thread so that read/write access issues are caught, and by adding infrastructure for testing the 'Insert into New File' action. The actions also now don't assume that they are invoked with read or write access, and instead obtain those permissions only when they are needed.

The Java version of this refactoring will be in a follow-up CL.

Test: KotlinActionProcessorTest
Bug: 280556462
Change-Id: I426b9d45274759b5681f5cdeffb11ec4916a387b