Show steps if they have problems that don't have locations
Filters that yielded XML elements (typically from the merged manifest)
were always hidden because those elements did not have a corresponding
code location.

To solve this, LocationProvider was generalized from a PsiElement
specific form (XML elements were turned into null LocationProviders) to
an interface with PsiElement and XML element specific implementations.

The PsiElementLocationProvider is simply the old LocationProvider
implementation. The new ElementLocationProvider for XML elements
currently has a null Location (since we still haven't figured out how to
tie them to code locations - but this way that's more explicit). The
main benefit here is that results from XML elements no longer silently
fall into the ether, but are instead explicitly treated as problems
without locations. Thus, the "does this step have problems" logic for
whether to show/hide it works as expected.

NB: This is sufficient for the Giraffe version b/c we only show/hide
steps and don't list any results. The remaining comment addresses the
results list active in Canary.

Of course, now we have LocationProviders that can only provide null
Locations. We already had handling for null Locations due to
SmartPsiElementPointer invalidation, so the list of result links will
elide them without issue. However, this could be confusing to a user
since it says to review problems without listing any.

Thus, we update the results message to tell the user about the number of
nonlocatable problems (if any exist) and to review the list of locatable
problems (if any exist). In practice, a filter is either one or the
other, but the logic handles a mixed case if it ever comes up.

Minor singluar/plural issues for that message were also fixed in a more
thorough manner.

Bug: 284349213
Test: N/A
Change-Id: I770f74012fc3e72a184606abadf1b1bc8f4ee5f5