Make URLs and long words break in messages
This CL adds a BreakMarkerInserter in charge of injecting break-markers
in long words and tokens, so that they can then be replace()d with any
desired string. This is used to allow URLs and other similar runs to
line break when used in the UI, by inserting a <wbr> tag in the HTML
rendered from Markdown.

This is needed because such long uninterrupted runs bork Swing's text
layout, making it unable to respect the width constraints and resulting
in clipped long text lines.

The new logic is used in the MarkdownConverter, but will also be needed
to allow reference URLs to wrap when we implement that section, until I
figure out how to custom-ellipsize them (if it is at all possible).

Bug: 280750424
Test: unit tests
Change-Id: I5007dcf37f9611af9757271c91b7e466bb9ba148