Make ComposePreviewView responsible for refreshing previews
The refresh mechanism shouldn't directly ask the surface to refresh the
preview elements. Instead, it should go through the ComposePreviewView,
who can then decide which elements will be refreshed. In Lite Mode, for
example, this can be only the preview element of the selected tab, while
in default mode it can be all the elements corresponding to the main
surface NlModels.

Bug: 279393638
Test: compose-designer tests should pass after this refactoring
Change-Id: Ic7459986ef80e4a69e59457b48e5900c8c9ae172