Autoupdate of 2 submodules
*tools/adt/idea updated to 12687aafe0d016145826077de92628d27c9e3edf
[NPW] Rewrite resolveDependency in terms of Dependency

Most of the uses are to generate a dependency identifier, so make it
explicit that we are working with Dependencies and RichVersions where
appropriate.

In the case of applying a plugin, however, we are not constructing a
Dependency directly (though we are using one using the special Gradle
plugin marker artifact); in that case, use the repositoryUrlManager
directly.

Bug: 279886738
Test: extended existing test
Change-Id: I685910c364811baaf349dc9577b0d42a8bfac7e4

*tools/base updated to 2dfa86add2110212be8c5e4d111afc444c403d23
Define a .require factory for RichVersion

Construct a singleton range with the given Version as the singleton.
Reduces the temptation to parse the version's toString(), which would
be lossy in edge cases.

Bug: 279886738
Test: added
Change-Id: I88decd55711f70e1d29eea81ecf4a90ac80053d7