API 28 - flag all uses of `Build.SERIAL`
This filter wasn't flagging the old `Build.SERIAL` usage if the required
manifest permission (`READ_PHONE_STATE`) for the new usage `Build.getSerial()`
was present. We want to flag both old usages (so they can be migrated to the
new usage) and if the new usage's required permission is missing.

Bug: 285579784
Test: N/A
Change-Id: I7c44071326fad7578cfcaa1d7199b2ff9c744b8f