Remove fake history injection
This CL reverts the change made by ag/23062261 which injects
a fake Q&A into the conversation history when it is empty.

With the 6/1 model update, this appears to no longer be necessary, after
testing against the same suite of questions which originally
justified the change - see go/studio-bot-prompting.

This also has the benefit of making our environment easily reproducible
outside of AS. We would like to stop relying on building and running Studio Bot to test prompts and response quality, and discrepancies like fake history make that more difficult.

This also appears to fix one of the issues in b/285390576, that the model sometimes replies with a generic greeting to a specific question at the beginning of a conversation.

Bug: 285390576
Test: manual
Change-Id: I5c4b905d077c34a3465e7abc137a9950b8b1f77b