Refactor dialog usages to be testable
This CL closes test coverage holes where the Kotlin and Java code
insertion processors require dialogs. Dialogs are used to choose names
for newly created files and functions. We now use a mock DialogProvider
to fill in these names with static default values. Some refactoring and
simplification in the areas where these dialogs are used is also
contained in this CL, which is necessary in order for the new tests to
pass and be readable.

Bug: n/a
Test: JavaInsertionProcessorTest, KotlinInsertionProcessorTest
Change-Id: I6aa2dd7ad3ae34e2bbc454bb32f7268a64249cee