Update UrlAssistantDatabaseConnection query and execute methonds.
Renamed underlying Connection to jdbcConn;
Added a transaction helper method to be able to execute a batch of SQLs;
Added lock using the singleton itself to sql execution methods to avoid concurrent read / write to the table as we are reusing the jdbcConn. In our use case, there shouldn't be concurrent read/write, and we don't care about the performance overhead of the lock.

Bug: 285938228
Test: No new tests yet. Keep existing tests pass.
Change-Id: I251b0994d6dc9995b185bfd7d76b377f3c7b5fbd