Autoupdate of 2 submodules
*tools/adt/idea updated to 56dba5e8dc3d9764a04f9bc8a4fa721401ab60b2
Replace uses of GradleCoordinate.parseVersionOnly

Attempt to be clear about whether the thing being parsed is a version
identifier or possibly a rich version specifier, and use the
appropriate parser.  We use the lowerBound even when the original code
uses COMPARE_PLUS_HIGHER because, if there is a prefix specifier, the
app compat migration should in fact replace the prefix with the base
version so that at least that base version is used.

Bug: 279886738
Test: existing
Change-Id: I540748f90b14848598136c3afbad92dd935dcf40

*tools/base updated to 8eb153cf59b1e1ea6f970ccf8f0019c08fecba95
Deprecate GradleCoordinate.parseVersionOnly

Bug: 279886738
Test: existing
Change-Id: Ic8d27f811bc4788e1eff27849f7a7b2e71d4a20f