Autoupdate of 2 submodules
*tools/adt/idea updated to 12114b343317ba6702844fdc7bffc2d7ec6ff3a6
Save theme and configuration with legacy snapshot

This will allow "Go To Declaration" to work on snapshots saved
from a legacy device.

Fixes: 285138114
Test: Adopted existing test
Change-Id: I5097bbdfadd3c06281de8bd8332852c96c83ee3d

*tools/base updated to 8d0b67d973646c951e68e6ae348b276e5896a097
Store the theme name instead of stringTable indices in AppContext

This will allow us to detect theme changes when an application
starts an activity with a different theme.

Fixes: 285350575
Test: Unit test added
Change-Id: Ic8f8cc0d3f1e63b5622cb664733287e0e53f9bb5