[CAP] Revise gcertstatus logic
It now parses the output instead of relying on the status code, since
the code will be 90 if either the LOAS2 or Corp SSO cookie is missing
(and only the latter is used by CAP).

This changelist also adds more debug logging for when the proxy server
fails to start.

Bug: N/A
Test: N/A
Change-Id: Id4c75928bee97ae04ead3455abfc4f6b27a27cb0