Autoupdate of 2 submodules
*tools/studio/google/cloud/tools updated to 8f1a0f5c72fd3b3a9c70c56101e1b229ddd08a46
Add LoginState rule

Test: n/a
Bug: 283753191
Change-Id: Ic1aff376b2a76990b65210a37415e223512e38db

*tools/vendor/google updated to 6c14c2b301fc055e2660a4807da1721102538ba9
Emit connections when logged in

This change makes CrashlyticsConfigurationManager emit connections every
time the user logs in. Previously, connections are emitted once (per
gradle sync), regardless of whether the user is logged in or not. So if
they are not, then AQI encounters permission errors when testing iAm
permission, which is shown when the user logs in.

Test: new CrashlyticsConfigurationManagerTest
Fixes: 283753191
Change-Id: Ic95b2eeccb1b1b8034583db4ff93541b8d565249