Disable 2256-checker-vector-replacement for SVE
The 2256-checker-vector-replacement test expects that the loop
optimization pass produces a VecLoad vector operation alongside a scalar
ArrayGet operation, which both reference the same heap object. An LSE
bug is tested when the LSE pass replaces the VecLoad with a scalar.

However, predicated vectorization via SVE eliminates the need for the
scalar tail loop, meaning that no ArrayGet is currently produced by the
loop optimzation pass in this case. Further, SVE currently does not
support LSE, meaning the bug would not be reproduced even if the test
was refactored to generate both an ArrayGet and a VecLoad.

Therefore this CL disables the test when SVE is enabled. When SVE
later supports LSE, the regression test can be appropriately refactored
and re-enabled.

Author: Richard Neill <richard.neill@arm.com>

Test: 2256-checker-vector-replacement (running with SVE enabled)

Change-Id: Iccc03935d02abbcf513a553ad810523494251918