Display citations in separate box
In order to do this, we needed not only to create the necessary UI
components and wire them up, but also to build the logic to parse and
persist them in the history.

A minor cosmetic issue is that, when switching themes, the links are not
automatically getting the new colour, even though we are forcing a hard
reset of the editor component that hosts them. I am kind of out of ideas
at this point for it; I am hoping nobody will notice, anyway :)

Fixes: 280219158
Fixes: 286053858
Test: manual
Change-Id: I13c8625d66b06fbceef382c4f5296863ad3d5877