Enable a flag for users to switch to new artifact handling
This adds an option in the Gradle expriemental settings to switch to
using the new variant reselection to obtain additional artifacts for
libraries.

Currently this API has a negative impact on the amount of memory used by
Gradle. However it does allow more correct handling in some cases which
may be worth the trade off.

Bug: 197636221
Test: N/A
Change-Id: Ia48fdc51fef1afb6267f183639fd9065a74c40fb