Fix target SDK project info with --project
When using `--project XML_FILE` and partial analysis, the min and target
SDK information was not being read from the manifest in the
`--report-only` phase. This can lead to many conditional incidents not
being reported or being incorrectly reported.

For `ManualProject`, we now skip the manifest reading initiated by
`LintDriver`, and instead read the merged manifest when creating the
`ManualProject`.

Fixes: 287438781
Test: enhanced existing unit test
Change-Id: I5634e893f38712800360b8a6791686e477782fa6