Autoupdate of 2 submodules
*tools/base updated to b0e1e58e55a716b1682926fe2cac7af9f5833946
Avoid linking project to AdbSession scope

invokeOnDisconnection is used to track the end of a device's ADB session
by waiting for its scope to terminate. However, the device's scope is
not a child of the project scope, since AdbSession is now
application-scoped. Thus, if the project scope terminates before the
device scope, the disconnection handler will retain the project in
memory.

Instead, we need to make sure that our project scope is a parent of any
job that runs on device scope. Thus, awaitTermination() is used within a
project-scoped coroutine to implement device disconnection handlers.

Bug: 275572248
Test: LeakChecker
Change-Id: Ie4297f96ac560840e0de79ea5101238e1a55ebaa

*tools/studio/google/cloud/testing updated to 7a562356e94592fbd54924d4b465a8cec9887789
Replace invokeOnDisconnection with awaitTermination

See change Ie4297f96ac560 for details.

Bug: 275572248
Test: LeakChecker
Change-Id: Icee9811a6993eeb013a488651de4be8830c838fb