Autoupdate of 2 submodules
*tools/adt/idea updated to 4d345ba633716623c93a4e464f2b92f7c176a0fd
Remove checks for missing AppContext and Configuration

We are removing the optimization in the agent that only sent the
AppContext and Configuration when they change. They are now always
present, so there is no need to check.

Test: NA
Bug: 287265444
Change-Id: Ice3b563e8bf2c3befbc44d5d8a73decd9718b549

*tools/base updated to 5ccf71cb71a5d2de925c85a27e8ec0f40b03d7aa
Always send AppContext and Configuration

This optimization is source of bugs and provides minimal performance
gains.

I tested the average size of an event from Layout Inspector on the
Sunflower app, using the embedded Layout Inspector. First with the
optimization ON, then OFF.

Average event size with optimization ON:  2114 bytes
Average event size with optimization OFF: 2229 bytes

Averages are calculated on ~300 messages in each configuration.

Test: NA
Bug: 287265444
Change-Id: Ib03d656b7dc404f8490e028f3f89ae1b167f63fe