Autoupdate of 2 submodules
*tools/adt/idea updated to f24293dc326824014212b93f39aa1b43dc3623bd
Update Lint baselines

ag/23647023 fixed an issue with AnnotationHandler which necessitated
changed to RestrictToDetector, and this resulted in some issues from baselines being fixed, and others having their descriptions changed.
So we update the baselines to reflect this.

Bug: 286595849, 287350230
Test: n/a
Change-Id: I087a2cccf6345dedd15e3201e6e89e1fb2f56368

*tools/base updated to c1decd8145c626c60e60f1d4fc27586814c4dff6
Handle type references in AnnotationHandler

...to match the spec of AnnotationUsageType.CLASS_REFERENCE

This CL also updates detectors which relied on the incorrect behavior.
In particular, this fix unearthed a problem with RestrictToDetector.
This detector wasn't handling the IntelliJ VisibleForTesting annotation
(see the description of b/287350230). When Lint was run on our own
codebase with the fixed AnnotationHandler, new errors were reported
which revealed this problem. In order for Lint to pass, this detector
also needs to be fixed.

Bug: 286595849, 287350230
Test: AnnotationHandlerTest, RestrictToDetectorTest
Change-Id: I96e7f9d67c255dc3db6cf104f3f7e410e0044c2f