Autoupdate of 2 submodules
*tools/adt/idea updated to c9de75cb0c9598e7c4518f53cba15e2fabcfe93d
Rearrange "More Actions" on Welcome Screen Project tab

Order them as requested.  Note that the ordering still depends on xml
processing order, in that "Profile or Debug APK" is specified to go
last, as is "SDK Manager".  This is unavoidable with the actions
declared in xml files, given that the separator has no id and so
cannot be referenced as part of the ordering.

Bug: 287418329
Test: manual
Change-Id: I99818b1510daac7753101d0426e65b77bed976dc

*tools/studio/google/cloud/tools updated to 03a247c64d4b6b186c14d39843523f67d0ee49d7
Reorganize "More Actions" on Welcome Screen Project tab

Make "Import an Android Code Sample" be first

Bug: 287418329
Test: manual
Change-Id: I2c293bf5418dc4006b645641f80fd16625c10f50