Use StartupActivity in new project wizard
This is necessary in order to ensure correct
ordering between project generation and Gradle sync
being triggered. Using ProjectActivity is not possible
right now as it does not provide ordering guarantees.

See https://youtrack.jetbrains.com/issue/IJPL-90 for details.

During the execution of activity, a modal task is launched
to ensure the generation happens in the background as source
generation logic requires that. This area should be cleaned
up more, and that's tracked in b/287942576.

Bug: 280359788
Test: //tools/adt/idea/android-uitests:NewComposeProjectTest
Change-Id: I9764df973b66136bbbfa9efc80da1a9a284e4a1a