Fix spurious highlighting in Version Catalog files
The scope used in GradleKtsVersionCatalogReferencesSearcher was too
wide, particularly in the case where the search is Local (e.g. "find
references within this file", as in the highlighting of usages and
declarations).  If the scope is local, do not expand it at all; if it
is a GlobalSearchScope, construct a KotlinScriptSearchScope that
delegates to that scope, so that we expand our search by the least
possible.

Bug: 287647449
Test: manual
Change-Id: Ic75d88dfe9a1c888eb76cf0dfd8001bba73e65b3