Better component attribution for instances of inner classes
Some integration e2e tests showed memory fluctuations between `loaders`
component and shared clusters. It turned out that a large chunk of
memory is held by an instance of an inner class
(ZipResourceFile$MyJarUrlStreamHandler) and the ownership of this class
fluctuates in a similar way. So the processing of inner classes was
improved, and we also added ZipResourceFile to the `loaders` component
declaration.

Bug: 285034032
Test: None
Change-Id: I93ca939e843b27f7d31d22b0fa66c6ccc6b55ee3