Implement blocked code snippet UI
There is now a new code blocked chunk type in town! It replaces code
chunks when they are blocked. The original code chunk is discarded and
can't be retrieved.

The old recitations UI and logic has been removed as it's unused and, if
it ever will be, it'll mean we have plenty of changes to make to it
anyway.

Also fixes issues when copying to the clipboard, where the <wbr>s are
treated as white space when copying. The fix is to replace() them all
with an empty string. The copy logic has been significantly simplified
and improved, also to better align with the new UX, as the result of
copying was now misaligned with the actual UI. This partially fixes the
b/286983519 issue, except when using the keyboard shortcut to copy. A
separate CL will address that scenario since it requires further
investigation.

Last minor fix is to hide the bot response bubble when showing an error
as it's not necessary, and align paddings correctly.

Fixes: 286019458
Bug: 286983519
Test: unit tests, visual inspection
Change-Id: I29f23527f67e21d7eabf34c4c1a2a0e92ffcf87b