Refactor scratch file handling and add test coverage
Extracts the scratch file handling into a separate file instead
of being done in ChunkAction.kt, and moves other necessary utils to a new file.

We also switch from VfsUtil.saveText to Document.insertString for adding text to scratch files. The former caused issues which were caught by the new tests.

A minor bug in KotlinInsertionProcessor is also fixed, where insertion behavior was incorrectly coupled to the currently open file, which may not be the same as the target file when inserting into a scratch file. This was also caught by the new tests.

Test: ScratchFileHandlerTest
Bug: 287358812
Change-Id: I055f21095f57002519270d1ce15b818c09452358