Remove total power rail row when power data empty
When there is no power rail data present in the system
trace capture, we should not display the row in the
summary panel for total power consumed. Currently, it
is shown with a value of 0 because no data is present,
but this is misleading as power is definitely consumed,
but we just don't have the data to say how much.

Fixes: 287290051
Test: added
Change-Id: I3d89afa552fc3a029722029c801601101f1bc507