Autoupdate of 2 submodules
*tools/adt/idea updated to 039c909c4768932775219de8056fd34afa7eaf8e
Remove obsolete "use filters in TSDKV" flag

Filters in general are now part of the stable assistant, though entirely
in the background to hide steps on startup, so we no longer need the
ability to turn them off.

Specific subfeatures of filters (run on start, WIP filters, redo
filters) remain in development and have their own flags.

Fixes: 288152071
Test: N/A
Change-Id: I94a23ecf2217cf474b99d8c6ca1467e4087051fb

*tools/vendor/google updated to 4e8af09312a9c253591a4bed055f3b9ea8e86b38
Remove usage of TSDKVUA_FILTERS flag from assistant

Filters in general are now part of the stable assistant so we no longer
need this toggle.

Bug: 288152071
Test: N/A
Change-Id: Iaaa1a0f81a2a20f3023d6a35520785fd05aaba5d