Autoupdate of 2 submodules
*tools/adt/idea updated to 395180edd5a34bf0f62a761a803663d6c52a04a0
Remove test for old offline device support

The tested functionality is being replaced, both the old Device File
Explorer tool window and the particular support for offline devices in
the Device Provisioner.

Bug: 287347772
Test: Removed
Change-Id: I2f60583896bee5f82104cb68e56311bcb0a625c7

*tools/base updated to d5251f9105b6e1076a165ccc8c97c5171b7b420a
Repace OfflineDeviceHandle with unclaimedDevices

Rather than representing offline devices as DeviceHandles, expose the
underlying ConnectedDevices to clients that need to see that there is an
connected but offline device.

Bug: 287347772
Test: Updated
Change-Id: I5432227cbc8853e7303cd8a82848728276815b5d