Autoupdate of 2 submodules
*tools/adt/idea updated to 8cedcf61985ceb8e3d53748502a1b4ca83cb2af0
Associate kmp android test compilations with the main compilation

JB is providing a fix for marking modules on the IDE side as test
modules based on whether a compilation has other compilations associated
with it.

Bug: 287449949
Test: snapshot tests
Change-Id: I6a23e673ac6c5e2778b84355ca79702f16ff4bf2

*tools/base updated to 1fb3623aa6a5b7c4b706b4c512032da4f05da764
Associate kmp android test compilations with the main compilation

JB is providing a fix for marking modules on the IDE side as test
modules based on whether a compilation has other compilations associated
with it.

Bug: 287449949
Test: snapshot tests
Change-Id: I9e455e3041e61d35d9bf3d1592aa7124a1abf19d