Maybe fix changeLanguageLevel on modules with unapplied plugins
We mustn't emit android-specific Language Level configuration (such as
`android.compileOptions { ... }`) if the module has a declaration of
an android-related plugin that is not applied.  Use the applied
plugins of the module rather than the declared plugins to detect
which (if any) of the configurations need to be included or updated on
a language level change.

Bug: 287809697
Test: none
Change-Id: Iee9692214d43ab634a68a300ce2267347eab1448