Update D8 R8 to 8.2.15-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.2.15-dev
Test: ./gradlew check
Bug:

Change-Id: I213568454b2b6645c596b3d4451cc55e6dd932f6