Disable assistant interactions when sync hasn't succeeded
If sync hasn't succeeded and a filter runs then it returns no results
and the step is treated as irrelevant (on first run) or completed (on
later runs). This is especially problematic on the first run b/c the
only way to reset is to restart Studio.

In order to guard against this failure mode, we want to prohibit filters
from running until sync has succeeded. There's two ways filters can run:
on assistant startup and on a verify button click. The solution is to
disable both of these when sync hasn't succeeded.

- New service to monitor sync updates
  - Listens for updates and caches latest sync result
  - Fires a stateful button refresh after an update
  - Exposes whether the sync succeeded or not, which is all the asst
    cares about
  - Threadsafe
- Assistant open action disabled when sync hasn't succeeded by querying
  the sync service
- Verify buttons are disabled when sync hasn't succeeded by querying the
  sync service
  - Update is triggered by the sync service when something changes

Fixes: 272710802
Test: N/A
Change-Id: Ic78f6e79e57a94846f6b84e133efdf2bc2350bbf