Add back notification for Essentials Mode
A recent change made the notification not appear properly due to
changing the registry value. Additionally, add the startUpActivity
to notify users of being in Essentials mode. This is fixed upstream,
but as the notification will be changing for Essentials Mode in 23.2
adding this in as a short-term fix to make users aware.

Fixes: 288323121
Change-Id: I8d860b9516a22f6beca8365590c2f190be923b42