Autoupdate of 2 submodules
*tools/adt/idea updated to 3659b6553fbb3e0ad8dc3ff35e92c73eee797a9e
Check if common sourceSet should be android lazily

Move the check that the common sourceSet should be treated as an android
sourceSet to be done only when the dependencies are resolved during
import action. This check was done during afterEvaluate of the kmp
android plugin, where allKotlinSourceSets property wasn't ready yet.

Bug: 287448185
Test: snapshot tests
Change-Id: If37b9bc75436116373f624fdb5530e6c32b476e0

*tools/base updated to 95fbcd0baa7fcf7f7f4b3599804de7d688b4e248
Use kotlin plugin default compilation associator

The object is internal, and JB is providing a public one soon.

Bug: 243387425
Test: existing
Change-Id: Id4819c5ba2003f264150ee8b108fe801670c5c55