Check if module has an AndroidFacet before adding resources
We are supposed to fall back on a module selection dialog if no suitable
module is found based on the open file when e.g. merging resources. But
there are some cases when a module is identified, but it isn't an
Android module, so when the action starts it throws an error. We should
instead eagerly determine if modules are suitable for Android
operations, and if not fall back on the dialog.

Tests for these actions and scenarios will be added in a follow-up CL.

Fixes: 287923325
Test: manual
Change-Id: I96edc1dd99514119f5e7b4ea5ddfed82aa09cb42