Handle multi-module projects when inserting XML/resources
Instead of defaulting to the first module in the list of modules
in the project, we now check if there is exactly one module, and
if not we let the user choose which one to insert a resource
into. Also extracts out shared code from Xml and resource insertion.

This is a follow-up to ag/23795418. Tests will be added soon.

Bug: n/a
Test: manual
Change-Id: I58c5582ab859d65bbf6adb4f9fc8ed0859ae3c16