Rename "goto test" to "gototest"
Because the command parser doesn't understand commands with multiple
words. If we add more similar commands, we can refactor them into
"/goto" and use "test" as an argument, instead of treating "goto test"
as a standalone command. But right now it doesn't make sense to
change the command to "goto" because only "goto test" is supported.

Fixes: 288844566
Test: manual
Change-Id: Iea30a20621f383c3b8f0a3bdea511cd48e321e29