Autoupdate of 2 submodules
*tools/adt/idea updated to 6a36ebf894e1a1af5884bb8599e0cff1ae53190f
Implement device details for AVDs

Recreates the existing AVD details to maintain feature parity. This is
slated for future replacement with something more curated and more
generic across device types.

Bug: 285599389
Test: Added (for the part expected to remain)
Change-Id: I393680827c50a3d478ef609937654746a2c49ed1

*tools/base updated to 2776f22f9f27b2525f372bf440043a242170408a
Add AvdInfo.properties to LocalEmulatorProperties

This is needed for the device details display, which is basically a
dump of this map.

Bug: 285599389
Test: updated
Change-Id: I0582e2e20642fe84937cc9818cff4e75d186773e