Reset state if closeConnection failed in DirectAccessConnection
To prevent duplicate calls, DirectAccessConnection sets its state to
DISCONNECTING when it is being closed. However, if an error occurs
during the close process, the state may not be updated to
DISCONNECTED. This can prevent the connection from being closed
properly and may cause problems if the connection is attempted to be
closed again.
This commit fixes this issue by always setting state to DISCONNECTED.
This commit also suspends connect method before adb connected.

Bug: 289142351
Test: updated
Change-Id: I0f88dfd969113b26c8c7a57c0b12b4bf6253f5e6