Add back missing margin between paragraphs
The issue is with the CSS styling for <p> tags in text chunks. It used
to work, then for whatever reason it stopped working (perhaps a platform
change? It's the only thing that changed). The rule is still there, and
should work, but something overrides it.

This CL changes a bit the way the rule is set since it gets overridden
otherwise. Now we're injecting it in a <style> tag int the <head> of any
TectChunk. This needs to happen again every time we append a chunk, too,
as we replace the document (we parse it with JSoup, add the needed nodes
and then re-apply it to the JEditorPane) and it'd get lost, but it is
not an expensive operation.

While we're at it, we also remove the bottom spacing from lists, since
there is always either another paragraph (with top margin) or another
chunk panel (with a vertical gap in between), so it's redundant. Now all
items have a consistent vertical spacing.

Fixes: 289180120
Test: visual inspection
Change-Id: I70997851431933f1c53d742b6ea5e33b47f88dc8