Forward clicks on buttons to row
Clicking on a row normally selects the row; however, if the row contains
a control that has mouse listeners, that control will process the click,
and the row will not receive it. To process the click, we need to add
another listener to the control that forwards it to the row.

Also, change the mouseClicked handler in CategoryTable to a mousePressed
handler, since that's how the ActionGroup popup works. (Without this,
you can do a small drag on the overflow button to launch the overflow
menu without selecting the row.)

Bug: 276973202
Test: added
Change-Id: Ic8a1d1668b4cba0695e6f9b17c3e874f670bd666