Remove black strips at the edges of the display image
Due to video size alignment requirements, video frame may have slightly different aspect ratio compared to the device display. When using the media projection API this causes appearance of narrow black strips at the edges of the device display image.

To address this problem we make sure in the agent that the video frame is taller rather than wider than the perfectly scaled display image. The video decoder in Studio removes the extra strips at the top and bottom of the display image.

Test: existing
Bug: 271874832
Change-Id: If44b531e03bbda7d633ae2b7e58072956df2eeec