Autoupdate of 2 submodules
*tools/adt/idea updated to 2fa924e01268cc9e55e366ba1d7ee656994a87fe
Use app:title if present for Preference in Component tree

Official documentation uses app:title and not android:title when
defining preferences. However, the Component tree only looked at
android:title for displaying. This makes sure to first try the app:title
attribute.
In addition, this assigns SwitchPreferenceHandler to be used for
SwitchPreferenceCompat as well as SwitchPreference.

Fixes: 177246254
Test: tests added
Change-Id: I4bda090da581b741b013f9a996edf8eddd2e4921

*tools/base updated to 979e39f4019264cab7c728adb0f4884780989688
Add SwitchPreferenceCompat constant

Bug: 177246254
Test: N/A
Change-Id: I23e517cf808f9a991e1b66d04f9e7325dcd41765