Autoupdate of 2 submodules
*tools/adt/idea updated to f5bec72a6a5f1115f0d54bfefa2375ddce876962
Improve RenderSecurityManager thread check

In order to decide when the RenderSecurityManager should apply its
policies we were only relying in setActive. This could introduce certain
cases where the security manager would prevent a legitimate thread from
doing an operation.
This ensures that the RenderSecurityManager only applies to the render
thread.

Fixes: 228523690
Test: //tools/adt/idea/rendering/...
Change-Id: Ie68c419a4310d4a168bfc022970a5bf73165fd85

*tools/base updated to 9286c8156b40e14075d5f3aab5abf02d70d80eed
Improve RenderSecurityManager thread check

In order to decide when the RenderSecurityManager should apply its
policies we were only relying in setActive. This could introduce certain
cases where the security manager would prevent a legitimate thread from
doing an operation.
This ensures that the RenderSecurityManager only applies to the render
thread.

Fixes: 228523690
Test: //tools/adt/idea/rendering/...
Change-Id: I7ac536378968edc904839a33c5a92f4bf66fc8ad