[ASfP] Optimize ASfP sync for building a large number of modules
Currently ASfP sync builds soong modules in a Java subprocess that
executes a bash command with inline args. When the number of modules
gets too large, the process fails because there is an upper limit to the
size of the supplied args. Storing these modules in temporary files and
leveraging command substitution helps circumvent this issue.

However, this is not the only problem; the `m` command also imposes a
constraint on the size of its args. Even though the bash command becomes
short after substituting modules with the aforementioned tmp files, it
eventually resolves to invoking `m` on a potentially long list of args.
A feasible workaround here is to split the modules into smaller, more
manageable chunks.

Bug: 289401083
Test: Manual
Change-Id: Ia8e2c795f68e934853a3780c1797f68680fe34ad