Move actions to BGT thread
All actions using findComposePreviewManagersForContext
should run in the BGT thread since the method might access the
VirtualFile from the context (which requires BGT).

Most of the actions can run on BGT without changes and were just in EDT
because obtaining the status of the Compose Preview required EDT in the
past. This is no longer the case and the actions can be moved to BGT.

Fixes: 289769734
Test: //tools/adt/idea/compose-designer/..
Change-Id: Iad600575277ecf5a4b1691f03c091a55e446d95e