Autoupdate of 3 submodules
*tools/adt/idea updated to f7b6695725145370e632b4102466fcaa08854be8
[Ps] Provide a refactoring for buildConfig

If the project is using a version of AGP where the default is true,
adjust module build files so that we can insert a property setting the
default to false.

If the project is using a version of AGP where the default is false,
perform any necessary module build file modifications and then remove
the property.

Bug: 262240431
Bug: 289219116
Test: manual, added
Change-Id: I1440eaa125f2e2ab97d954d20391260badb72310

*tools/analytics-library updated to 43948a781c7808b7e5d392c75909b2e3421e3b28
Update studio_stats.proto

Logs-Approval: cl/545216432
Bug: 262240431
Test: existing
Change-Id: I72015f5aa2ec5fb09f28e6bcb0331012d8e365a5

*tools/base updated to dabbf4f210a1ae7ddbb5de1c1040d8f6826620ab
Include plugin descriptors in AGP artifacts

Bug: none
Test: none
Change-Id: I00338ea598fea7fa4db1adb82c4fea4c5cf979bf