Autoupdate of 2 submodules
*tools/adt/idea updated to bc4516ad72a254b0de9ceee60ad506b3ef8ae386
Run JDK-related Old Agp Tests

The tests were annotated with agpVersions = "7.4.0", but AGP_7_4 is
currently associated with version 7.4.1.  Prior to that change,
AGP_7_4 was associated with version 7.4.0-alpha09, so in neither case
did the agpVersion filter allow the test to actually run.

Bug: none filed
Test: this is it
Change-Id: I778f710bcaa669f4daac947f63b9e2ba1385f494

*tools/base updated to 510b5f386bc45ad6a6932161a1d882092a883eb5
Remove 7.4.0 (really 7.4.0-alpha09) previous version

Nothing should use it any more.  If anything does that is not caught
by presubmit, it should use 7.4.1 instead.

Bug: none filed
Test: existing
Change-Id: I28a8f9b7bf7fe332029e92b9dffa14d2f01af369