[Ps] Remove one parse/unparse round trip
There's no particular need to parse the data from an
UnresolvedDependencies Sync Issue into a GradleCoordinate, only for
the one user to directly convert it back into a String (and then
attempt to match it with strings generated from a substantially
different process).

The "show in project structure" hyperlink may appear in marginally
more cases, and in some of those cases the link will work; even when
it doesn't completely work, however, its presence is not too unhelpful
as it will at least open the PSD on the declared dependencies tab of
the applicable module.

Bug: 279886738
Test: existing test adjusted
Change-Id: If06a547f6b6509926a3e5ee1e484630f0b4f9dcc