Autoupdate of 2 submodules
*tools/adt/idea updated to a06e0d1b3e101f65aed3edabbb4813a94dc37706
Allowed paths should be defined by the ScreenshotEnvironmentContext

The default rendering paths use PathManager and do not work for
ScreenshotEnvironmentContext. This change allows the caller to pass the
paths so ScreenshotEnvironmentContext can do the right thing.

Test: N/A
Bug: N/A
Change-Id: I83692fcbffda54e96e193f32cc68be38520bca42

*tools/base updated to e53edf8ff69e14ee995c55d546bbf109fe036982
Allowed paths should be defined by the ScreenshotEnvironmentContext

The default rendering paths use PathManager and do not work for
ScreenshotEnvironmentContext. This change allows the caller to pass the
paths so ScreenshotEnvironmentContext can do the right thing.

Test: N/A
Bug: N/A
Change-Id: I13b0558fbe657dedbe452f87a4c4ebde877a3c5e