Update static_assert dependent on kPageSize
This patch is part of the chain preparing for making kPageSize
non-constexpr in a future patch.

In most of the assertions the kPageSize is replaced with kMinPageSize.
However, in the MemMapTest_Reservation the assertion is replaced with
ASSERT_LT instead: while similarly to the other cases the input values
of the static_assert are dependent on the kPageSize, they can't be
replaced with kMinPageSize or other constant expressions.

Test: Same as for I5430741a8494b340ed7fd2d8692c41a59ad9c530.
      The whole patches chain was tested as a whole.
Change-Id: I18efc71052273ef1f4d13948cbccbf8bd6379911