Merge hide property into Showability
Fundamentally, the hide and show properties are mutually exclusive both
at the individual annotation level (a single annotation can only be a
hide or a show annotation not both), and at the `Item` level, an `Item`
cannot be shown and hidden from the same API surface.

Previously, they were separate properties (`hide` and `showability`).
This change removes the `hide` property and modifies the existing
`Showability` properties `show` and `recursive` so use the new
`ShowOrHide` type which is a 3-state enum that can indicate whether
an `Item` is annotated with a show or hide annotation or neither.

When an `Item` has both show and hide annotations (there are tests that
cover that) then it will resolve in favor of the show which is what the
previous behavior did but was not very clear.

This purposely does not change `Item.hasHideAnnotation()` to return
`showability.hide()` as `Item.hasHideAnnotation()` is used to compute
`PsiItem.originalHidden` and does not consider show annotations but
`showability.hide()` does. Follow up changes will clean that up.

Bug: 292484829
Test: ./gradlew
Change-Id: I718621cca3d648f74fbf61695cfedcaaaea53458