libavb_rs: add partition GUID callback
Adds support for the `get_unique_guid_for_partition()` user op callback.

This one is a bit trickier because this callback isn't always used, but
if it is used, implementation is a lot nicer if we pull in the `uuid`
Rust library.

Rather than force all users to pull in `uuid` whether they want to use
this functionality or not, we hide it behind a `uuid` feature flag so
that we only require the dependency if it's actually used.

Bug: b/290110273
Test: atest
Change-Id: I29dcfb750bdeb665f9eb1f11ba3e4eb0afc044c5